requirements.original.txt
# Core packages
fastapi==0.110.2
uvicorn==0.29.0
redis==5.2.1
psutil==5.9.8

# Azure & security
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
cryptography==44.0.0
PyJWT==2.10.1

# Pinecone
pinecone-client==3.0.0

# NLP / spaCy
spacy==3.7.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# Miscellaneous
annotated-types==0.7.0
async-timeout==5.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1

# LangChain ecosystem – let pip resolve versions

langchain-core
langchain
langchain-openai
langchain-community
langchain-text-splitters
langsmith
openai
tenacity
